# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Link API",
        "description": "API documentation for Link",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Default server"
        }
    ],
    "paths": {
        "/api/register": {
            "post": {
                "tags": ["Auth"],
                "summary": "Register a new user",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": ["name", "email", "password"],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ada Lovelace"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "ada@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": ["Auth"],
                "summary": "Login and receive API token",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": ["email", "password"],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "ada@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "secret123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "422": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/user": {
            "delete": {
                "tags": ["Auth"],
                "summary": "Delete the authenticated user",
                "operationId": "3c7a2fa0e79dfdc62f5cab7ebca9d825",
                "responses": {
                    "200": {
                        "description": "Account deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/matrix": {
            "get": {
                "tags": ["Matrix"],
                "summary": "List matrices for the authenticated user",
                "description": "Display a listing of the resource.",
                "operationId": "010519e26c36a4006a11537a3cd8c48c",
                "parameters": [
                    {
                        "name": "order",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": ["asc", "desc"]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Matrix list"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": ["Matrix"],
                "summary": "Create a matrix",
                "description": "Store a newly created resource in storage.",
                "operationId": "2bd0553de6ff0f17adf4b573cfa2609a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": ["grid", "size"],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "My Matrix"
                                    },
                                    "grid": {
                                        "type": "string",
                                        "example": "1 2 3\n4 5 6\n7 8 9"
                                    },
                                    "size": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Matrix created"
                    },
                    "400": {
                        "description": "Invalid grid"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/matrix/{matrix}": {
            "get": {
                "tags": ["Matrix"],
                "summary": "Get a matrix by id",
                "description": "Display the specified resource.",
                "operationId": "eea85c481dac42545833278230a1dc84",
                "parameters": [
                    {
                        "name": "matrix",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Matrix details"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": ["Matrix"],
                "summary": "Update a matrix",
                "description": "Update the specified resource in storage.",
                "operationId": "ac7e1d9b02913f1ea62b53c1a394b54f",
                "parameters": [
                    {
                        "name": "matrix",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Renamed Matrix"
                                    },
                                    "grid": {
                                        "type": "string",
                                        "example": "1 2\\n3 4"
                                    },
                                    "size": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Matrix updated"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": ["Matrix"],
                "summary": "Delete a matrix",
                "description": "Remove the specified resource from storage.",
                "operationId": "727935dd2e8cb5986f89740b900b2c0c",
                "parameters": [
                    {
                        "name": "matrix",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Matrix deleted"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/matrix/{matrix}/calculate": {
            "get": {
                "tags": ["Matrix"],
                "summary": "Calculate visibility for a matrix",
                "description": "Calculate the visibility of the matrix",
                "operationId": "029995fd76f98ac49c0050070dae2476",
                "parameters": [
                    {
                        "name": "matrix",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Visibility result"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/": {
            "get": {
                "tags": ["Web"],
                "summary": "Welcome page",
                "operationId": "5ed014c185ff373f284d57420dfc7d7d",
                "responses": {
                    "200": {
                        "description": "Welcome HTML"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use the API token as: Bearer {token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Auth",
            "description": "Auth"
        },
        {
            "name": "Matrix",
            "description": "Matrix"
        },
        {
            "name": "Web",
            "description": "Web"
        }
    ]
}